Switch show & episode parsing to expat w/improvements.
I changed the calls that return a show and/or (an) episode(s) to use the
expat parser.  In my testing of the full Jeopardy! data (about 2.4M worth
of show + episodes) the old code took about 1.8 seconds, while the new
took 0.6 seconds (both were parsing a local, uncompressed xml file).  I
also switched the get_show_and_episodes() call to download the zip data,
which will save a lot of bandwidth (e.g. that 2.4M of Jeopardy! data is
only 77K, for instance).  To add the latter feature I needed a language
abbreviation, so I added the ability to specify the language to the
TheTVDB constructor (which defaults to "en").