suppress myconnpy bug which errors on sql warnings
This dirty hack turns off warning notification globally on the db until turned back on with ```SET sql_notes =1;```

The alternative is to find each place in code which executes a SQL statement and wrap it in a try..except blcok. This solution is more performant on limited platforms (like routers).